+++ src/components/home/HomeCardCarouselLazy.tsx
@@ -8,6 +8,7 @@ const HomeCardCarousel = dynamic(
       default: m.HomeCardCarousel,
     })),
   {
+    ssr: false,
     loading: () => (
       <div
         className="mx-auto h-[230px] w-full max-w-[min(327px,100%)] overflow-hidden"